Fix kinsol dummy wrapper
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@19178 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/solver/nonlinearSolverNewton.c
@@ -430,7 +430,7 @@ static int _omc_newton(integer* n, double *x, double *fvec, double* eps, double*
 
   while(*info >= 0)
   {
-    debugStreamPrint(LOG_NLS_V, 0, "**** start Iteration: %d  *****", *maxfev-l);
+    debugStreamPrint(LOG_NLS_V, 0, "**** start Iteration: %d  *****", (int) *maxfev-l);
     /* calculate the function values */
     (*f)(n, x, fvec, &iflag, userdata,currentSys);
     (*nfev)++;
